Update README.md
# LocalVillager

This plugin allows you to spawn a universal villager. You can give it a name in the configuration file and it can also display random messages.
This plugin is aimed at working in small servers, for entertainment purposes.

# Installing

- Unpack the archive;
- Put it in the folder with the plugins;
- Start the server;
- Stop the server;
- Finish.

# Command

/reloading - reload config file
/local - spawn custom villager

# Config file

Here you can set the name, as well as specify messages that will be displayed randomly.

# Permissions

local.spawn - Allows spawning villager

# License

Use of this software is governed by the zlib/libpng License.